tsanv3: revert Tid/StackID

+++ compiler-rt/lib/asan/asan_debugging.cpp
@@ -55,13 +55,11 @@ uptr AsanGetStack(uptr addr, uptr *trace, u32 size, u32 *thread_id,
   if (alloc_stack) {
     if (chunk.AllocTid() == kInvalidTid) return 0;
     stack = chunk.GetAllocStack();
-    if (thread_id)
-      *thread_id = static_cast<u32>(chunk.AllocTid());
+    if (thread_id) *thread_id = chunk.AllocTid();
   } else {
     if (chunk.FreeTid() == kInvalidTid) return 0;
     stack = chunk.GetFreeStack();
-    if (thread_id)
-      *thread_id = static_cast<u32>(chunk.FreeTid());
+    if (thread_id) *thread_id = chunk.FreeTid();
   }
 
   if (trace && size) {
